requirements.txt
chromadb==1.5.5
groq==1.1.2
pytest==9.0.2
python-dotenv==1.2.2
rank_bm25==0.2.2
sentence_transformers==5.3.0
